Update plots.R
Proposition to fix plot_world & plot_world_data to handle when the number of data points is null in some countries. "breaks" is now an argument in the function "plot_world" (default =NULL) and I add conditions to define the colors correctly. The breaks parameter is calculated in plot_world_data. This change should not affect the other cases handled by the function.