add_pop function
I used dplyr package in the new function "add_pop" I thought it was necessary to add it to the required package in this code. I don't know if I need to call the package here or in the R script of the function itself.
I added an R script related to data manipulation, with for now a function that allows to add population data, summing the population of the different years in case a study lasts for several years.