Update plot_world and plot_world _data
The function plot_world now has an additional parameter which takes a dataset with the disease-free countries. Those are plotted on the map in a different category. the function plot_world_data has got an update about the breaks, legend labels and colors such that the zero is appearing as first in the legend.